+++ src/test/resources/com/pestphp/pest/SimpleHigherOrderNotTest.php
@@ -1,0 +1,3 @@
+<?php
+
+it('lives')->not->assertTrue(true);